@@ -23,14 +23,42 @@ namespace LightQueryProfiler.Shared.Services
 
             foreach (ProfilerEvent readEvent in readEvents)
             {
-                // ignore events created by itself
-                if (!readEvent.GetEventKey().Contains("LightQueryProfiler", StringComparison.OrdinalIgnoreCase))
+                // Ignore events created by the profiler itself
+                if (IsProfilerGeneratedEvent(readEvent))
                 {
-                    newEvents.Add(readEvent);
+                    continue;
                 }
+
+                newEvents.Add(readEvent);
             }
 
             return newEvents;
+        }
+
+        private static bool IsProfilerGeneratedEvent(ProfilerEvent profilerEvent)
+        {
+            // Check client_app_name in Actions
+            if (profilerEvent.Actions?.TryGetValue("client_app_name", out var appName) == true
+                && appName?.ToString()?.Contains("LightQueryProfiler", StringComparison.OrdinalIgnoreCase) == true)
+            {
+                return true;
+            }
+
+            // Check batch_text in Fields (for sql_batch_completed events)
+            if (profilerEvent.Fields?.TryGetValue("batch_text", out var batchText) == true
+                && batchText?.ToString()?.Contains("LightQueryProfiler", StringComparison.OrdinalIgnoreCase) == true)
+            {
+                return true;
+            }
+
+            // Check statement in Fields (for rpc_completed events)
+            if (profilerEvent.Fields?.TryGetValue("statement", out var statement) == true
+                && statement?.ToString()?.Contains("LightQueryProfiler", StringComparison.OrdinalIgnoreCase) == true)
+            {
+                return true;
+            }
+
+            return false;
         }
 
         public void PauseProfiling(string sessionName)
